// src/shared/lib/validation/enums.ts
enum Auth {
  REQUIRED_MESSAGE = "Це поле є обов'язковим",
  INVALID_MESSAGE_EMAIL = "Неправильний формат електронної адреси",
  INVALID_MESSAGE_PASSWORD = "Неправильний формат паролю",
  MIN_LENGTH_EMAIL = 1,
  MIN_LENGTH_PASSWORD = 8,
}

export default Auth;